+++ package.json
@@ -64,6 +64,7 @@
     "prettier-plugin-organize-imports": "^4.3.0",
     "reload": "^3.4.2",
     "stylelint": "^16.24.0",
-    "stylelint-config-standard": "^39.0.0"
+    "stylelint-config-standard": "^39.0.0",
+    "vitest": "^3.2.4"
   }
 }
