+++ src/components/topFilm/topFilm.js
@@ -1,6 +1,6 @@
-import { formatDuration } from '../../helpers/durationFormatHelper.js'
-import { formatRating } from '../../helpers/ratingFormatHelper.js'
-import { getRatingType } from '../../helpers/ratingTypeHelper.js'
+import { formatDuration } from '../../helpers/durationFormatHelper/durationFormatHelper.js'
+import { formatRating } from '../../helpers/ratingFormatHelper/ratingFormatHelper.js'
+import { getRatingType } from '../../helpers/ratingTypeHelper/ratingTypeHelper.js'
 import Component from '../core/baseComponent.js'
 
 export default class TopFilm extends Component {
