+++ src/helpers/columnCountHelper/columnCountHelper.js
