+++ src/helpers/ratingFormatHelper/ratingFormatHelper.js
